callbacks -> icqcallbacks and split callbacks into separate files
